decoupled app and config
